[Admin Analytics] refactor frontend
