+++ client/web/src/nav/GlobalNavbar.tsx
@@ -35,7 +35,6 @@ import {
     ButtonLink,
     PopoverTrigger,
     useWindowSize,
-    Badge,
 } from '@sourcegraph/wildcard'
 
 import { AuthenticatedUser } from '../auth'
@@ -143,9 +142,6 @@ const AnalyticsNavItem: React.FunctionComponent = () => {
         <NavAction className="d-none d-sm-flex">
             <Link to="/site-admin/analytics/search" className={classNames('font-weight-medium', styles.link)}>
                 Analytics
-                <Badge className="mx-1" variant="merged">
-                    Experimental
-                </Badge>
             </Link>
         </NavAction>
     )
