+++ client/web/src/site-admin/analytics/AnalyticsCodeIntelPage/index.module.scss
@@ -1,0 +1,3 @@
+.border {
+    border-bottom: 1px solid var(--border-color-2);
+}
