+++ client/web/src/site-admin/analytics/AnalyticsComingSoonPage/index.module.scss
@@ -1,0 +1,4 @@
+.large-icon {
+    width: 6rem !important;
+    height: 6rem !important;
+}
