+++ client/web/src/site-admin/analytics/AnalyticsSearchPage/queries.ts
@@ -1,0 +1,39 @@
+import { gql } from '@sourcegraph/http-client'
+
+const analyticsStatItemFragment = gql`
+    fragment AnalyticsStatItemFragment on AnalyticsStatItem {
+        nodes {
+            date
+            count
+            uniqueUsers
+        }
+        summary {
+            totalCount
+            totalUniqueUsers
+        }
+    }
+`
+
+export const SEARCH_STATISTICS = gql`
+    query SearchStatistics($dateRange: AnalyticsDateRange!) {
+        site {
+            analytics {
+                search(dateRange: $dateRange) {
+                    searches {
+                        ...AnalyticsStatItemFragment
+                    }
+                    resultClicks {
+                        ...AnalyticsStatItemFragment
+                    }
+                    fileViews {
+                        ...AnalyticsStatItemFragment
+                    }
+                    fileOpens {
+                        ...AnalyticsStatItemFragment
+                    }
+                }
+            }
+        }
+    }
+    ${analyticsStatItemFragment}
+`
