+++ client/web/src/site-admin/analytics/components/ValueLegendList.tsx
@@ -5,9 +5,9 @@ import classNames from 'classnames'
 
 import { Text } from '@sourcegraph/wildcard'
 
-import { formatNumber } from '../lib'
+import { formatNumber } from '../utils'
 
-import styles from '../index.module.scss'
+import styles from './index.module.scss'
 
 interface ValueLegendItemProps {
     color: string
