+++ client/web/src/site-admin/analytics/components/index.module.scss
@@ -33,17 +33,8 @@
     margin-top: 1rem;
 }
 
-.large-icon {
-    width: 6rem !important;
-    height: 6rem !important;
-}
-
 .text-wrap {
     white-space: pre-wrap;
-}
-
-.border {
-    border-bottom: 1px solid var(--border-color-2);
 }
 
 .count-box-value {
