--- client/web/src/site-admin/analytics/index.tsx
@@ -1,4 +1,0 @@
-export { AnalyticsSearchPage } from './AnalyticsSearchPage'
-export { AnalyticsNotebooksPage } from './AnalyticsNotebooksPage'
-export { AnalyticsUsersPage } from './AnalyticsUsersPage'
-export { AnalyticsCodeIntelPage } from './AnalyticsCodeIntelPage'
