--- client/web/src/site-admin/analytics/lib/index.ts
@@ -1,1 +1,0 @@
-export const formatNumber = (value: number): string => Intl.NumberFormat('en', { notation: 'compact' }).format(value)
